Updated with the btree algorithm
One thing to note, in line 6 of the book, we check for north and east,
and chose either one to link. Somehow, it does not work on this code, I
have to use South and East. Will have to check why I can't follow the
code as is given in the book